+++ lib/src/state/settings_by_app.rs
@@ -1,19 +1,21 @@
 use regex::Regex;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use serde_alias::serde_alias;
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all(deserialize = "snake_case"))]
+#[serde(rename_all = "snake_case")]
 pub enum AppExtraFlag {
+    /// Start the app in the center of the screen as floating in the wm.
     Float,
+    /// Force manage this app in the wm.
     Force,
+    /// Unmanage this app in the wm.
     Unmanage,
+    /// Pin this app in all the virtual desktops in the wm.
     Pinned,
-    // only for backwards compatibility
-    ObjectNameChange,
-    Layered,
-    BorderOverflow,
-    TrayAndMultiWindow,
+    /// Hide this app on the dock/taskbar.
+    Hidden,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
@@ -45,7 +47,9 @@ pub enum MatchingStrategy {
     Legacy,
 }
 
+#[serde_alias(SnakeCase)]
 #[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "camelCase")]
 pub struct AppIdentifier {
     pub id: String,
     pub kind: AppIdentifierType,
@@ -124,15 +128,24 @@ impl AppIdentifier {
     }
 }
 
+#[serde_alias(SnakeCase)]
 #[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "camelCase")]
 pub struct AppConfig {
+    /// name of the app
     pub name: String,
+    /// category to group the app under
     pub category: Option<String>,
-    pub bound_monitor_idx: Option<usize>,
-    pub bound_workspace_name: Option<String>,
+    /// monitor index that the app should be bound to
+    pub bound_monitor: Option<usize>,
+    /// workspace name that the app should be bound to
+    pub bound_workspace: Option<String>,
+    /// app identifier
     pub identifier: AppIdentifier,
+    /// extra specific options/settings for the app
     #[serde(default)]
     pub options: Vec<AppExtraFlag>,
+    /// is this config bundled with seelen ui.
     #[serde(default)]
     pub is_bundled: bool,
 }