+++ src/apps/settings/modules/shared/store/app/StateBridge.ts
@@ -1,10 +1,10 @@
 import { UserSettings } from '../../../../../../shared.interfaces';
-import { parseAsCamel, VariableConvention } from '../../../../../shared/schemas';
+import { parseAsCamel } from '../../../../../shared/schemas';
 import { IdWithIdentifierSchema } from '../../../../../shared/schemas/AppsConfigurations';
 import { ISettings } from '../../../../../shared/schemas/Settings';
 import { pick } from 'lodash';
 
-import { AppConfiguration, ApplicationOptions } from '../../../appsConfigurations/domain';
+import { AppConfiguration } from '../../../appsConfigurations/domain';
 import { RootState } from '../domain';
 
 export const YamlToState_Apps = (yaml: anyObject[]): AppConfiguration[] => {
@@ -20,13 +20,7 @@ export const YamlToState_Apps = (yaml: anyObject[]): AppConfiguration[] => {
         workspace: ymlApp.bound_workspace || null,
         identifier: parseAsCamel(IdWithIdentifierSchema, ymlApp.identifier),
         isBundled: ymlApp.is_bundled || false,
-        // options
-        [ApplicationOptions.Float]: ymlApp.options?.includes(ApplicationOptions.Float) || false,
-        [ApplicationOptions.Unmanage]:
-          ymlApp.options?.includes(ApplicationOptions.Unmanage) || false,
-        [ApplicationOptions.Pinned]: ymlApp.options?.includes(ApplicationOptions.Pinned) || false,
-        [ApplicationOptions.ForceManage]:
-          ymlApp.options?.includes(ApplicationOptions.ForceManage) || false,
+        options: ymlApp.options,
       });
     }
   });
@@ -34,10 +28,7 @@ export const YamlToState_Apps = (yaml: anyObject[]): AppConfiguration[] => {
   return apps;
 };
 
-export const StaticSettingsToState = (
-  userSettings: UserSettings,
-  state: RootState,
-): RootState => {
+export const StaticSettingsToState = (userSettings: UserSettings, state: RootState): RootState => {
   const { jsonSettings, yamlSettings, themes, layouts, placeholders, wallpaper } = userSettings;
 
   return {
@@ -47,7 +38,7 @@ export const StaticSettingsToState = (
     availableThemes: themes,
     availableLayouts: layouts,
     availablePlaceholders: placeholders,
-    appsConfigurations: YamlToState_Apps(yamlSettings),
+    appsConfigurations: yamlSettings,
   };
 };
 
@@ -63,25 +54,4 @@ export const StateToJsonSettings = (state: RootState): ISettings => {
     'devTools',
     'language',
   ]);
-};
-
-export const StateAppsToYamlApps = (
-  appsConfigurations: AppConfiguration[],
-  template?: boolean,
-): anyObject[] => {
-  return appsConfigurations
-    .filter((appConfig) => !appConfig.isBundled)
-    .map((appConfig: AppConfiguration) => {
-      const options = Object.values(ApplicationOptions).filter((option) => appConfig[option]);
-      const yamlApp = {
-        name: appConfig.name,
-        template: template || undefined,
-        category: appConfig.category || undefined,
-        bound_monitor: appConfig.monitor ?? undefined,
-        bound_workspace: appConfig.workspace || undefined,
-        identifier: VariableConvention.fromCamelToSnake(appConfig.identifier),
-        options: options.length ? options : undefined,
-      };
-      return yamlApp;
-    });
 };